feat: 문제 정보를 포함한 학생 상태 조회 기능 작성
-성 각 서비스계층에서 데이터를 조회해 합쳐서 반환
- ContestService: 진입 가능한 대회를 조회하는 책임
- TrackingStudentService: 대회 중 학생의 문제 정보를 가져오는 책임
- ProblemService: 문제를 몇번 풀었는지를 확인하는 책임

+++ src/main/java/com/twoori/contest_server/domain/student/vo/ProblemStatus.java
@@ -1,4 +1,4 @@
 package com.twoori.contest_server.domain.student.vo;
 
-public record ProblemStatus(Long problemID, Long contentId, Long countOfTry) {
+public record ProblemStatus(Long problemID, Long contentId, Integer countOfTry) {
 }
